+++ scripts/host/install_php.sh
@@ -1,10 +1,13 @@
 #!/usr/bin/env bash
 
-vagrant_dir=$(cd "$(dirname "$0")/../.."; pwd)
-host_os=$(bash "${vagrant_dir}/scripts/host/get_host_os.sh")
+cd "$(dirname "${BASH_SOURCE[0]}")/../.." && vagrant_dir=$PWD
 
-# Enable trace printing and exit on the first error
-set -ex
+source "${vagrant_dir}/scripts/output_functions.sh"
+
+status "Installing PHP"
+incrementNestingLevel
+
+host_os="$(bash "${vagrant_dir}/scripts/host/get_host_os.sh")"
 
 if [[ ${host_os} == "Windows" ]]; then
     curl http://windows.php.net/downloads/releases/archives/php-5.6.9-nts-Win32-VC11-x86.zip -o "${vagrant_dir}/lib/php.zip"
@@ -16,14 +19,13 @@ if [[ ${host_os} == "Windows" ]]; then
     rm -rf "${vagrant_dir}/lib/php/*.back"
 fi
 
-php_executable=$(bash "${vagrant_dir}/scripts/host/get_path_to_php.sh")
+php_executable="$(bash "${vagrant_dir}/scripts/host/get_path_to_php.sh")"
 if ! ${php_executable} -v | grep -q 'Copyright' ; then
-    set +x
-    echo "Automatic PHP installation is not available for your host OS. Please install any version of PHP to allow Magento dependencies management using Composer. Check out http://php.net/manual/en/install.php"
-    exit 255
-    set -x
+    error "Automatic PHP installation is not available for your host OS. Please install any version of PHP to allow Magento dependencies management using Composer. Check out http://php.net/manual/en/install.php"
+    decrementNestingLevel
+    exit 1
 else
-    set +x
-    echo "PHP installed successfully."
-    set -x
+    success "PHP installed successfully"
 fi
+
+decrementNestingLevel