+++ scripts/host/relink_sample_data.sh
@@ -1,16 +1,18 @@
 #!/usr/bin/env bash
 
-vagrant_dir=$(cd "$(dirname "$0")/../.."; pwd)
+cd "$(dirname "${BASH_SOURCE[0]}")/../.." && vagrant_dir=$PWD
+
+source "${vagrant_dir}/scripts/output_functions.sh"
+
 magento_ce_dir="${vagrant_dir}/magento2ce"
 magento_ee_dir="${magento_ce_dir}/magento2ee"
 magento_ce_sample_data_dir="${magento_ce_dir}/magento2ce-sample-data"
 magento_ee_sample_data_dir="${magento_ce_dir}/magento2ee-sample-data"
 php_executable="$(bash "${vagrant_dir}/scripts/host/get_path_to_php.sh")"
-install_sample_data=$(bash "${vagrant_dir}/scripts/get_config_value.sh" "magento_install_sample_data")
+install_sample_data="$(bash "${vagrant_dir}/scripts/get_config_value.sh" "magento_install_sample_data")"
 
-
-# Enable trace printing and exit on the first error
-set -ex
+status "Linking/unlinking sample data according to config.yaml"
+incrementNestingLevel
 
 install_ee=0
 if [[ -f "${magento_ce_dir}/app/etc/enterprise/di.xml" ]]; then
@@ -19,53 +21,48 @@ fi
 
 # As a precondition, disable CE sample data
 if [[ -f "${magento_ce_sample_data_dir}/dev/tools/build-sample-data.php" ]]; then
-    ${php_executable} -f "${magento_ce_sample_data_dir}/dev/tools/build-sample-data.php" -- --command=unlink --ce-source="${magento_ce_dir}" --sample-data-source="${magento_ce_sample_data_dir}" --exclude=true
-    set +x
-    echo "CE Sample data disabled"
-    set -x
+    status "Unlinking CE sample data"
+    ${php_executable} -f "${magento_ce_sample_data_dir}/dev/tools/build-sample-data.php" -- --command=unlink --ce-source="${magento_ce_dir}" --sample-data-source="${magento_ce_sample_data_dir}" --exclude=true 2> >(logError) > >(log)
 fi
 # As a precondition, disable EE sample data
 if [[ -f "${magento_ee_sample_data_dir}/dev/tools/build-sample-data.php" ]]; then
-    "${php_executable}" -f "${magento_ee_sample_data_dir}/dev/tools/build-sample-data.php" -- --command=unlink --ce-source="${magento_ce_dir}" --sample-data-source="${magento_ee_sample_data_dir}" --exclude=true
-    set +x
-    echo "EE Sample data disabled"
-    set -x
+    "${php_executable}" -f "${magento_ee_sample_data_dir}/dev/tools/build-sample-data.php" -- --command=unlink --ce-source="${magento_ce_dir}" --sample-data-source="${magento_ee_sample_data_dir}" --exclude=true 2> >(logError) > >(log)
+    status "Unlinking EE sample data"
 fi
 
 if [[ ${install_ee} -eq 1 ]]; then
-    "${php_executable}" -f "${magento_ee_dir}/dev/tools/build-ee.php" -- --command=link --ee-source="${magento_ee_dir}" --ce-source="${magento_ce_dir}" --exclude=true
+    status "Linking EE to CE"
+    "${php_executable}" -f "${magento_ee_dir}/dev/tools/build-ee.php" -- --command=link --ee-source="${magento_ee_dir}" --ce-source="${magento_ce_dir}" --exclude=true 2> >(logError) > >(log)
 fi
 
 if [[ ${install_sample_data} -eq 1 ]]; then
     # Installing CE or EE, in both cases CE sample data should be linked
     if [[ ! -f "${magento_ce_sample_data_dir}/dev/tools/build-sample-data.php" ]]; then
         # Sample data not available and should be enabled
-        set +x
-        echo "CE Sample data repository is not available. Recreate project using \"init_project.sh -fc\", which will delete Magento code base and recreate project from scratch. Or clone sample data to <project_dir>/magento2ce/magento2ce-sample-data."
-        set -x
-        exit 0
+        error "CE Sample data repository is not available. Recreate project using \"init_project.sh -fc\", which will delete Magento code base and recreate project from scratch. Or clone sample data to ${magento_ce_sample_data_dir}"
+        decrementNestingLevel
+        exit 1
     else
         # Sample data available and should be enabled
-        set +x
-        echo "CE Sample data enabled"
-        set -x
-        "${php_executable}" -f "${magento_ce_sample_data_dir}/dev/tools/build-sample-data.php" -- --command=link --ce-source="${magento_ce_dir}" --sample-data-source="${magento_ce_sample_data_dir}" --exclude=true
+        success "CE Sample data enabled"
+        cd ${magento_ce_sample_data_dir} && git fetch && git pull 2> >(logError) > >(log)
+        "${php_executable}" -f "${magento_ce_sample_data_dir}/dev/tools/build-sample-data.php" -- --command=link --ce-source="${magento_ce_dir}" --sample-data-source="${magento_ce_sample_data_dir}" --exclude=true 2> >(logError) > >(log)
     fi
 
     if [[ ${install_ee} -eq 1 ]]; then
         # Installing EE
         if [[ ! -f "${magento_ee_sample_data_dir}/dev/tools/build-sample-data.php" ]]; then
             # Sample data not available and should be enabled
-            set +x
-            echo "EE Sample data repository is not available. Recreate project using \"init_project.sh -fc\", which will delete Magento code base and recreate project from scratch. Or clone sample data to <project_dir>/magento2ce/magento2ee-sample-data."
-            set -x
-            exit 0
+            error "EE Sample data repository is not available. Recreate project using \"init_project.sh -fc\", which will delete Magento code base and recreate project from scratch. Or clone sample data to ${magento_ee_sample_data_dir}."
+            decrementNestingLevel
+            exit 1
         else
             # Sample data available and should be enabled
-            set +x
-            echo "EE Sample data enabled"
-            set -x
-            "${php_executable}" -f "${magento_ee_sample_data_dir}/dev/tools/build-sample-data.php" -- --command=link --ce-source="${magento_ce_dir}" --sample-data-source="${magento_ee_sample_data_dir}" --exclude=true
+            success "EE Sample data enabled"
+            cd ${magento_ee_sample_data_dir} && git fetch && git pull 2> >(logError) > >(log)
+            "${php_executable}" -f "${magento_ee_sample_data_dir}/dev/tools/build-sample-data.php" -- --command=link --ce-source="${magento_ce_dir}" --sample-data-source="${magento_ee_sample_data_dir}" --exclude=true 2> >(logError) > >(log)
         fi
     fi
 fi
+
+decrementNestingLevel