+++ scripts/provision/configure_environment.sh
@@ -1,8 +1,5 @@
 #!/usr/bin/env bash
 
-# Enable trace printing and exit on the first error
-set -ex
-
 use_nfs_for_synced_folders=$1
 guest_magento_dir=$2
 magento_host_name=$3
@@ -12,7 +9,12 @@ is_windows_host=$6
 
 vagrant_dir="/vagrant"
 
-# Enable Magento virtual host
+source "${vagrant_dir}/scripts/output_functions.sh"
+
+status "Configuring environment"
+incrementNestingLevel
+
+status "Enabling Magento virtual host"
 custom_virtual_host_config="${vagrant_dir}/etc/magento2_virtual_host.conf"
 default_virtual_host_config="${vagrant_dir}/etc/magento2_virtual_host.conf.dist"
 if [[ -f ${custom_virtual_host_config} ]]; then
@@ -24,27 +26,27 @@ enabled_virtual_host_config="/etc/apache2/sites-available/magento2.conf"
 cp "${virtual_host_config}" "${enabled_virtual_host_config}"
 sed -i "s|<host>|${magento_host_name}|g" "${enabled_virtual_host_config}"
 sed -i "s|<guest_magento_dir>|${guest_magento_dir}|g" "${enabled_virtual_host_config}"
-a2ensite magento2.conf
+a2ensite magento2.conf 2> >(logError) > >(log)
 
-# Disable default virtual host
-sudo a2dissite 000-default
+status "Disabling default virtual host"
+sudo a2dissite 000-default 2> >(logError) > >(log)
 
-# Configure composer
+status "Configuring composer"
 composer_auth_json="${vagrant_dir}/etc/composer/auth.json"
 if [[ -f ${composer_auth_json} ]]; then
-    set +x
-    echo "Installing composer OAuth tokens from ${composer_auth_json}..."
-    set -x
+    status "Installing composer OAuth tokens from ${composer_auth_json}"
     if [[ ! -d /home/vagrant/.composer ]] ; then
-      sudo -H -u vagrant bash -c 'mkdir /home/vagrant/.composer'
+        sudo -H -u vagrant bash -c 'mkdir /home/vagrant/.composer' 2> >(logError) > >(log)
     fi
     if [[ -f ${composer_auth_json} ]]; then
         cp "${composer_auth_json}" "/home/vagrant/.composer/auth.json"
     fi
 fi
 
-# Set permissions to allow Magento codebase upload by Vagrant provision script
 if [[ ${use_nfs_for_synced_folders} -eq 0 ]]; then
+    status "Setting permissions to allow Magento codebase upload by Vagrant provision script"
     chown -R vagrant:vagrant /var/www
     chmod -R 755 /var/www
 fi
+
+decrementNestingLevel