+++ scripts/provision/configure_environment_recurring.sh
@@ -20,35 +20,39 @@ function process_php_config () {
 }
 
 function init_php56 () {
-        add-apt-repository ppa:ondrej/php
-        apt-get update
-        apt-get install -y php5.6 php-xdebug php5.6-xml php5.6-mcrypt php5.6-curl php5.6-cli php5.6-mysql php5.6-gd php5.6-intl php5.6-bcmath php5.6-mbstring php5.6-soap php5.6-zip libapache2-mod-php5.6
-        echo '
-        xdebug.max_nesting_level=200
-        xdebug.remote_enable=1
-        xdebug.remote_host=192.168.10.1' >> /etc/php/5.6/mods-available/xdebug.ini
+    status "Installing PHP 5.6"
+    add-apt-repository ppa:ondrej/php 2> >(logError) > >(log)
+    apt-get update 2> >(logError) > >(log)
+    apt-get install -y php5.6 php-xdebug php5.6-xml php5.6-mcrypt php5.6-curl php5.6-cli php5.6-mysql php5.6-gd php5.6-intl php5.6-bcmath php5.6-mbstring php5.6-soap php5.6-zip libapache2-mod-php5.6 2> >(logError) > >(log)
+    echo '
+    xdebug.max_nesting_level=200
+    xdebug.remote_enable=1
+    xdebug.remote_host=192.168.10.1' >> /etc/php/5.6/mods-available/xdebug.ini
 }
 
 function isServiceAvailable() {
-    if service --status-all | grep -Fq ${1}; then
+    all_services="$(service --status-all 2> >(log))"
+    if [[ ${all_services} =~ ${1} ]]; then
         echo 1
     else
         echo 0
     fi
 }
 
-# Enable trace printing and exit on the first error
-set +x
-
 guest_magento_dir=$2
 use_php7=$4
 vagrant_dir="/vagrant"
 
-# Remove configs from host in case of force stop of virtual machine before linking restored ones
+source "${vagrant_dir}/scripts/output_functions.sh"
+
+status "Configuring environment (recurring)"
+incrementNestingLevel
+
+status "Removing configs from host in case of force stop of virtual machine before linking restored ones"
 cd "${vagrant_dir}/etc" && mv guest/.gitignore guest_gitignore.back && rm -rf guest && mkdir guest && mv guest_gitignore.back guest/.gitignore
 bash "${vagrant_dir}/scripts/guest/link_configs"
 
-# Make sure configs are restored on system halt and during reboot
+status "Making sure configs are restored on system halt and during reboot"
 rm -f /etc/init.d/unlink-configs
 cp "${vagrant_dir}/scripts/guest/unlink_configs" /etc/init.d/unlink-configs
 if [[ ! -f /etc/rc0.d/K04-unlink-configs ]]; then
@@ -61,12 +65,12 @@ if [[ ! -f /etc/rc0.d/K04-unlink-configs ]]; then
     ln -s /etc/init.d/unlink-configs /etc/rc6.d/K04-unlink-configs
 fi
 
-# Upgrade existing environment
+status "Upgrading existing environment"
 if [[ -f "${vagrant_dir}/.idea/deployment.xml" ]]; then
     sed -i.back "s|magento2ce/var/generation|magento2ce/var|g" "${vagrant_dir}/.idea/deployment.xml"
 fi
 
-# Copy varnish vcl file
+status "Copying varnish vcl file"
 custom_vcl_config="${vagrant_dir}/etc/magento2_default_varnish.vcl"
 default_vcl_config="${vagrant_dir}/etc/magento2_default_varnish.vcl.dist"
 if [ -f ${custom_vcl_config} ]; then
@@ -75,44 +79,44 @@ else
     cp ${default_vcl_config}  /etc/varnish/default.vcl
 fi
 
-# Setup PHP
+status "Setting up PHP"
 php_ini_paths=( /etc/php/7.0/cli/php.ini /etc/php/5.6/cli/php.ini )
 process_php_config ${php_ini_paths}
 
 if [[ ${use_php7} -eq 1 ]]; then
+    status "Configuring PHP 7"
     update-alternatives --set php /usr/bin/php7.0
     if [[ -d "/etc/php/5.6" ]]; then
-        a2dismod php5.6
+        a2dismod php5.6 2> >(logError) > >(log)
     fi
     sed -i "s|xdebug.remote_connect_back=1|xdebug.remote_host=192.168.10.1|g" /etc/php/7.0/cli/conf.d/20-xdebug.ini
-    a2enmod php7.0
+    a2enmod php7.0 2> >(logError) > >(log)
     # TODO: Fix for a bug, should be removed in 3.0
     sed -i "/zend_extension=.*so/d" /etc/php/7.0/cli/conf.d/20-xdebug.ini
     echo "zend_extension=xdebug.so" >> /etc/php/7.0/cli/conf.d/20-xdebug.ini
 else
+    status "Configuring PHP 5.6"
     if [[ ! -d "/etc/php/5.6" ]]; then
         init_php56
     fi
-    update-alternatives --set php /usr/bin/php5.6 && a2dismod php7.0 && a2enmod php5.6
+    update-alternatives --set php /usr/bin/php5.6 && a2dismod php7.0 && a2enmod php5.6 2> >(logError) > >(log)
     rm -rf /etc/php/5.6/apache2
     ln -s /etc/php/5.6/cli /etc/php/5.6/apache2
     sed -i "s|xdebug.remote_connect_back=1|xdebug.remote_host=192.168.10.1|g" /etc/php/5.6/mods-available/xdebug.ini
 fi
-service apache2 restart
-#end Setup PHP
+service apache2 restart 2> >(logError) > >(log)
 
-# Set up elastic search
 is_elastic_search_installed="$(isServiceAvailable elasticsearch)"
 if [[ ${is_elastic_search_installed} -eq 0 ]]; then
-    apt-get update
-    apt-get install -y openjdk-7-jre
-    wget https://download.elastic.co/elasticsearch/elasticsearch/elasticsearch-1.7.2.deb
-    dpkg -i elasticsearch-1.7.2.deb
-    update-rc.d elasticsearch defaults
+    status "Setting up ElasticSearch"
+    apt-get update 2> >(logError) > >(log)
+    apt-get install -y openjdk-7-jre 2> >(logError) > >(log)
+    wget https://download.elastic.co/elasticsearch/elasticsearch/elasticsearch-1.7.2.deb 2> >(logError) > >(log)
+    dpkg -i elasticsearch-1.7.2.deb 2> >(logError) > >(log)
+    update-rc.d elasticsearch defaults 2> >(logError) > >(log)
 fi
-# End set up elastic search
 
-# Enable email logging
+status "Enabling email logging"
 if [[ ${use_php7} -eq 1 ]]; then
     php_ini_file="/etc/php/7.0/cli/php.ini"
 else
@@ -122,5 +126,7 @@ pattern=";sendmail_path"
 php_config_content="$(cat ${php_ini_file})"
 if [[ ${php_config_content} =~ ${pattern} ]]; then
     sed -i "s|;sendmail_path =|sendmail_path = \"/vagrant/scripts/guest/log_email ${vagrant_dir}/log/email\"|g" ${php_ini_file}
-    service apache2 restart
+    service apache2 restart 2> >(logError) > >(log)
 fi
+
+decrementNestingLevel