+++ scripts/provision/export_env_variables_recurring.sh
@@ -1,6 +1,5 @@
 #!/usr/bin/env bash
 
-# Enable trace printing and exit on the first error
 set -e
 
 guest_magento_dir=$2
@@ -8,6 +7,11 @@ host_magento_dir=$5
 is_windows_host=$6
 host_vagrant_dir=$7
 vagrant_dir="/vagrant"
+
+source "${vagrant_dir}/scripts/output_functions.sh"
+
+status "Exporting environment variables"
+incrementNestingLevel
 
 if ! cat /etc/profile | grep -q 'export PATH=' ; then
     echo "export PATH=\$PATH:${vagrant_dir}/scripts/guest:$\"{guest_magento_dir}/bin\"" >> /etc/profile
@@ -23,3 +27,5 @@ if ! cat /etc/profile | grep -q 'export VAGRANT_ROOT=' ; then
     echo "export VAGRANT_ROOT=${vagrant_dir}" >> /etc/profile
     echo "export VAGRANT_ROOT_HOST=\"${host_vagrant_dir}\"" >> /etc/profile
 fi
+
+decrementNestingLevel